pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.5",
    "curl-cffi>=0.15.0",
    "fastapi>=0.135.3",
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "fng-api>=0.0.5",
    "pysocks>=1.7.1",
    "python-socketio[client]>=5.16.1",
    "requests[socks]>=2.33.1",
    "uvicorn[standard]>=0.44.0",
    "websocket-client>=1.9.0",
]
